+++ lua/gitlab/config.lua
@@ -3,7 +3,6 @@ local M = {}
 M.options = {
   glab_binary = "glab",
   gitlab_token_env = "GITLAB_TOKEN",
-  default_branch = "main",
   ci_file = ".gitlab-ci.yml",
   picker = "vim_ui",
   scratch_height = 15,